refactor: pull threads into their own module
RunSimThread is also now a stoppable thread.
